pyproject.toml
[project]
name = "data-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.6",
    "google-cloud>=0.34.0",
    "google-cloud-storage>=3.2.0",
    "kubernetes>=33.1.0",
    "pandas>=2.3.1",
    "planetary-computer>=1.0.0",
    "psycopg2-binary>=2.9.10",
    "pypgstac[psycopg]>=0.9.8",
    "pystac-client>=0.9.0",
    "rasterio>=1.4.3",
    "rio-cogeo>=5.4.2",
    "rio-tiler>=7.8.1",
]


[tool.setuptools.packages.find]
where = ["."]